Reorganize file trees
